requirements.txt
pySigma>=0.11.0
pySigma-backend-microsoft365defender>=0.3.0
PyYAML>=6.0
